+++ client/src/providerView.js
@@ -2,17 +2,17 @@ const vscode = require('vscode')
 const Provider = require('./provider')
 const Folder = require('./folder')
 
+
 class ProviderView {
 
-    // View Folder
-    static Container = class Container {
-
-        constructor(label, description, elements) {
+    static Item = class Item {
+        constructor(label, description, data) {
             this.label = label
             this.description = description
-            this.elements = elements
+            this.data = data
         }
     }
+
 
     // Provider View Class
     static #instance
@@ -32,8 +32,7 @@ class ProviderView {
     }
 
 
-
-
+    // change events
     #onDidChangeTreeData = new vscode.EventEmitter()
     get onDidChangeTreeData() {
         return this.#onDidChangeTreeData.event
@@ -56,7 +55,14 @@ class ProviderView {
         const Expanded = vscode.TreeItemCollapsibleState.Expanded
         const None = vscode.TreeItemCollapsibleState.None
 
-        if (element instanceof Provider) {
+        if (element instanceof ProviderView.Item) {
+            const item = new vscode.TreeItem(element.label, None)
+            item.description = element.description
+            item.contextValue = "itemContext"
+
+            return item
+        }
+        else if (element instanceof Provider) {
             const item = new vscode.TreeItem(element.identifier, None)
             item.contextValue = "providerContext"
             if (element.definedVersion)
@@ -71,15 +77,12 @@ class ProviderView {
             return item
         }
         else if (element instanceof Folder) {
-            const item = new vscode.TreeItem(null, Collapsed)
-            item.description = element.properties.description ?? element.name
-            item.contextValue = "folderContext"
+            const item = new vscode.TreeItem(element.properties.label, Collapsed)
+            if (null != element.properties.description)
+                item.description = element.properties.description.length > 0 ? `\\${element.properties.description}` : element.properties.description
+            else
+                item.description = element.name
 
-            return item
-        }
-        else if (element instanceof ProviderView.Container) {
-            const item = new vscode.TreeItem(element.label, Expanded)
-            item.description = element.description
             item.contextValue = "folderContext"
 
             return item
@@ -90,43 +93,11 @@ class ProviderView {
     async getChildren(element) {
 
         if (null == element) {
-
-            const requiredProvidersAtPath = await this.#client.sendRequest('requiredprovider.get')
-            
-            const rootFolder = new Folder()
-            for (const [path, providers] of Object.entries(requiredProvidersAtPath)) {
-                const reducedPath = vscode.workspace.workspaceFolders.reduce((path, folder) => path.replace(folder.uri.fsPath, ''), path)
-
-                let folderElements = []
-                for (const requiredProvider of Object.values(providers)) {
-                    let providerInfo = await this.#client.sendRequest('provider.info', requiredProvider.source)
-                    providerInfo = new Provider(providerInfo, this.#client)
-                    providerInfo.definedVersion = requiredProvider.version
-                    folderElements.push(providerInfo)
-                }
-
-                rootFolder.add(reducedPath, folderElements, {
-                    description: reducedPath
-                })
-            }
-
-            const foldersStart = []
-            const findSubFolders = current => {
-                if (current.content.length > 0)
-                    foldersStart.push(current)
-                else
-                    current.folders.forEach(v => findSubFolders(v))
-
-                return foldersStart
-            }
-
-            return findSubFolders(rootFolder).map(folder => folder.name.length > 1 ? folder : folder.content).flat()
+            return await this.getRootFolderElemets()
         }
-        else if (element instanceof ProviderView.Container) {
-            return await Promise.resolve(element.elements)
-        }
+
         else if (element instanceof Folder) {
-            return element.content //.concat(element.folders)
+            return await this.getChildItemForFolderContent(element.content)
         }
 
         return null
@@ -136,6 +107,94 @@ class ProviderView {
         // fire event from onDidChangeTreeData Event Emitter.
         this.#onDidChangeTreeData.fire()
     }
+
+
+
+
+
+
+    // returns topmost required_providers of configurations in all workspaces
+    async getRootFolderElemets() {
+
+        const terraformBlockAtPath = await this.#client.sendRequest('requiredprovider.get')
+
+        const rootFolder = new Folder()
+        for (const [initialPath, terraform] of Object.entries(terraformBlockAtPath)) {
+
+            const workspaceFolder = vscode.workspace.workspaceFolders.filter(folder => initialPath.includes(folder.uri.fsPath)).at(0)
+            if (null == workspaceFolder) continue
+
+            let pathArray = workspaceFolder.uri.fsPath.split('\\')
+            pathArray = pathArray.slice(0, pathArray.length - 1)
+            const reducedPath = initialPath.replace(pathArray.join('\\'), '')
+
+            // When multiple folders in workspace, shows the foldername as treeitem name 
+            if (vscode.workspace.workspaceFolders.length > 1) {
+                const pathArray = reducedPath.split('\\')
+                rootFolder.add(reducedPath, terraform, {
+                    label: pathArray[1],
+                    description: pathArray.slice(2).join('\\')
+                })
+            }
+            // When only one folder, omitts treeitem name
+            else {
+                rootFolder.add(reducedPath, terraform, {
+                    label: null,
+                    description: reducedPath.split(`\\`).slice(2).join('\\')
+                })
+            }
+        }
+
+        // Find topmost folders in workspaces
+        const foldersStart = []
+        const findSubFolders = current => {
+            if (current.content.length > 0)
+                foldersStart.push(current)
+            else
+                current.folders.forEach(v => findSubFolders(v))
+
+            return foldersStart
+        }
+
+        let processedFolders = []
+        for (const folder of findSubFolders(rootFolder)) {
+            if (folder.properties.label?.length > 0 || folder.properties.description?.length > 0)
+                processedFolders.push(folder)
+            else {
+                const childContent = await this.getChildItemForFolderContent(folder.content)
+                processedFolders = [...childContent, ...processedFolders]
+            }
+        }
+
+        return processedFolders
+    }
+
+
+    // returns terraform version and providers at that path
+    async getChildItemForFolderContent(content) {
+
+        let providerItems = Object.values(content[0].requiredProviders)
+        // Get provider data from api/cache and create instances of Provider-Class
+        providerItems = providerItems.map(requiredProvider =>
+            this.#client.sendRequest('provider.info', requiredProvider.source)
+                .then(providerInfo => {
+                    const provider = new Provider(providerInfo, this.#client)
+                    provider.definedVersion = requiredProvider.version
+                    return provider
+                })
+        )
+        providerItems = await Promise.all(providerItems)
+
+        // Item for showing required terraform version
+        if (content[0].requiredVersion) {
+            const item = new ProviderView.Item('terraform', content[0].requiredVersion)
+            providerItems = [item, ...providerItems]
+        }
+
+        return providerItems//.concat(element.folders)
+    }
+
+
 }