+++ CHANGELOG.md
@@ -4,32 +4,53 @@ All notable changes to the "terraform-quick-docs" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.21]
+
+### Improvements
+
+- removed obsolete Settings
+- optimised providerview feature
+- show required terraform version on providerview
 
 ## [0.0.20]
 
-- Add view for providers in current configuration
+### Feature
 
+- Add view for providers in current configuration
+  
 ## [0.0.19]
+
+### Fixed
 
 - fix null being ignored when parsin providers
 
 ## [0.0.18]
 
+### Improvements
+
 - Support hover for built-in resource: `terraform_data`
 
 ## [0.0.17]
+
+### Feature
 
 - optional support for other module sources
 
 ## [0.0.16]
 
+### Fixed
+
 - fixed error with certain module sources
 
 ## [0.0.15]
 
+### Fixed
+
 - fixed error with settings
 
 ## [0.0.14]
+
+### Feature
 
 - Queries terraform 'required_providers'-block for source and version
 - Providers that are defined there, will be found regardles if they are partner or official providers
@@ -39,54 +60,80 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [0.0.13]
 
+### Improvements
+
 - Tweaks to the function documentations
 
 ## [0.0.12]
+
+### Feature
 
 - Added hover support for terraform functions
 - Added commands to open terraform functions docs and addtional docs
 
 ## [0.0.11]
 
+### Improvements
+
 - Add filecaching on API-Calls
 
 ## [0.0.10]
 
+### Feature
+
 - Add Command for opening any Resource Documentation
 
 ## [0.0.9]
+
+### Improvements
 
 - Change getProviders API to use v2 with more options
 - Read providers list from static JSON, instead of calling API on each start of language server.
 
 ## [0.0.8]
 
+### Fixed
+
 - Fix errors with previous implementation
 
 ## [0.0.7]
+
+### Feature
 
 - Open Documentation for inline elements, such as 'data.<data_source>', '<resource_identifer>.<resource_name>'
 
 ## [0.0.6]
 
+### Feature
+
 - Settings to add support for more than the default Providers
 
 ## [0.0.5]
+
+### Feature
 
 - Add functionality for opening modules in the terraform registry
 
 ## [0.0.4]
 
+### Feature
+
 - Add a simple icon
 
 ## [0.0.3]
 
-- fix: terraform api doesn't return azure/azapi as a verified provider
+### Fixed
+
+- terraform api doesn't return azure/azapi as a verified provider
 
 ## [0.0.2]
+
+### Fixed
 
 - Update Changelog and Readme to released
 
 ## [0.0.1]
+
+### Feature
 
 - Initial release with basic functionality